Create a extension-base package
