+++ packages/extension-base/README.md
@@ -1,0 +1,3 @@
+# @polkadot/extension
+
+A signing extension
