+++ packages/extension-base/src/background/RequestBytesSign.ts
