+++ packages/extension-base/src/background/handlers/Extension.test.ts
@@ -2,13 +2,14 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import { ResponseAccountExport } from '@polkadot/extension/background/types';
+import { ResponseAccountExport } from '../types';
 
-import Extension from '@polkadot/extension/background/handlers/Extension';
-import State from '@polkadot/extension/background/handlers/State';
 import keyring from '@polkadot/ui-keyring';
 import ExtensionStore from '@polkadot/ui-keyring/stores/Extension';
 import { cryptoWaitReady } from '@polkadot/util-crypto';
+
+import Extension from './Extension';
+import State from './State';
 
 describe('Extension', () => {
   async function createExtension (): Promise<Extension> {
