+++ packages/extension-base/src/background/handlers/Extension.ts
