+++ packages/extension-base/src/background/handlers/State.ts
