+++ packages/extension-base/src/background/handlers/index.ts
