+++ packages/extension-base/src/background/handlers/subscriptions.ts
