+++ packages/extension-base/src/background/types.ts
