+++ packages/extension-base/src/defaults.ts
