+++ packages/extension-base/src/page/Injected.ts
