+++ packages/extension/package.json
@@ -5,14 +5,11 @@
   "author": "Jaco Greeff <jacogr@gmail.com>",
   "license": "Apache-2",
   "dependencies": {
-    "@polkadot/api": "^1.7.0-beta.10",
-    "@polkadot/extension-ui": "0.23.0-beta.1",
-    "@polkadot/keyring": "^2.6.2",
-    "@polkadot/ui-keyring": "^0.52.0-beta.7",
-    "extensionizer": "^1.0.1"
+    "@babel/runtime": "^7.8.7",
+    "@polkadot/extension-base": "0.23.0-beta.1",
+    "@polkadot/extension-ui": "0.23.0-beta.1"
   },
   "devDependencies": {
-    "@babel/runtime": "^7.8.7",
     "@polkadot/dev": "^0.50.31",
     "@types/chrome": "^0.0.100",
     "@types/firefox-webext-browser": "^70.0.1",
