+++ tsconfig.json
@@ -10,6 +10,8 @@
     "paths": {
       "@polkadot/extension": ["packages/extension/src"],
       "@polkadot/extension/*": ["packages/extension/src/*"],
+      "@polkadot/extension-base": ["packages/extension-base/src"],
+      "@polkadot/extension-base/*": ["packages/extension-base/src/*"],
       "@polkadot/extension-chains": ["packages/extension-chains/src"],
       "@polkadot/extension-chains/*": ["packages/extension-chains/src/*"],
       "@polkadot/extension-dapp": ["packages/extension-dapp/src"],
